tests: split off helpers from test_lnpeer into new lnhelpers.py
